+++ packages/google_maps_flutter/google_maps_flutter_android/example/integration_test/google_maps_tests.dart
@@ -2009,6 +2009,7 @@ Marker _copyMarkerWithClusterManagerId(
     position: marker.position,
     rotation: marker.rotation,
     visible: marker.visible,
+    // The deprecated parameter is used here to avoid losing precision.
     zIndex: marker.zIndex,
     onTap: marker.onTap,
     onDragStart: marker.onDragStart,